Upgrade to create-react-app 1.0.10
Run type checks on a separate process in browser dev for faster updates

+++ scripts/init.js
@@ -29,11 +29,8 @@ module.exports = function(
   originalDirectory,
   template
 ) {
-  const ownPackageName = require(path.join(
-    __dirname,
-    '..',
-    'package.json'
-  )).name;
+  const ownPackageName = require(path.join(__dirname, '..', 'package.json'))
+    .name;
   const ownPath = path.join(appPath, 'node_modules', ownPackageName);
   const appPackage = require(path.join(appPath, 'package.json'));
   const useYarn = fs.existsSync(path.join(appPath, 'yarn.lock'));
@@ -113,23 +110,25 @@ module.exports = function(
       }
     }
   );
-  fs.move(
-    path.join(appPath, 'npmignore'),
-    path.join(appPath, '.npmignore'),
-    [],
-    err => {
-      if (err) {
-        // Append if there's already a `.npmignore` file there
-        if (err.code === 'EEXIST') {
-          const data = fs.readFileSync(path.join(appPath, 'npmignore'));
-          fs.appendFileSync(path.join(appPath, '.npmignore'), data);
-          fs.unlinkSync(path.join(appPath, 'npmignore'));
-        } else {
-          throw err;
+  if (fs.existsSync(path.join(appPath, 'npmignore'))) {
+    fs.move(
+      path.join(appPath, 'npmignore'),
+      path.join(appPath, '.npmignore'),
+      [],
+      err => {
+        if (err) {
+          // Append if there's already a `.npmignore` file there
+          if (err.code === 'EEXIST') {
+            const data = fs.readFileSync(path.join(appPath, 'npmignore'));
+            fs.appendFileSync(path.join(appPath, '.npmignore'), data);
+            fs.unlinkSync(path.join(appPath, 'npmignore'));
+          } else {
+            throw err;
+          }
         }
       }
-    }
-  );
+    );
+  }
 
   let command;
   let args;
@@ -232,6 +231,8 @@ module.exports = function(
 function isReactInstalled(appPackage) {
   const dependencies = appPackage.dependencies || {};
 
-  return typeof dependencies.react !== 'undefined' &&
-    typeof dependencies['react-dom'] !== 'undefined';
+  return (
+    typeof dependencies.react !== 'undefined' &&
+    typeof dependencies['react-dom'] !== 'undefined'
+  );
 }
